Ajout gestion transactions

+++ src/BankDataAccess/Transactionobligatoire.cs
@@ -8,7 +8,7 @@ namespace BankDataAccess
         public int Idtransac { get; set; }
         public int Idcompte { get; set; }
         public string Nomtransaction { get; set; } = null!;
-        public double Montant { get; set; }
+        public decimal Montant { get; set; }
         public int Typeid { get; set; }
         public int Jour { get; set; }
